Configure webpack. Make modules
