+++ .gitignore
@@ -10,6 +10,7 @@ Thumbs.db
 # Вспомогательные файлы
 *.log*
 node_modules/
+*.pdf*
 
 # Папка с собранными файлами проекта
 build/
