+++ src/make-tasks.js
@@ -1,0 +1,269 @@
+import {default as getRandomNumber} from './utils.js';
+
+const taskNode = document.querySelector(`.board__tasks`);
+
+export default function showTasks(quantity) {
+  let taskElementString = ``;
+  const fragment = document.createDocumentFragment();
+  const parser = new DOMParser();
+  for (let i = 0; i < quantity; i++) {
+    taskElementString += templateTask;
+    const cardElementObj = parser.parseFromString(taskElementString, `text/html`);
+    fragment.appendChild(cardElementObj.body.childNodes[0]);
+  }
+  taskNode.appendChild(fragment);
+}
+
+export const onFilterLoaderClick = () => {
+  taskNode.innerHTML = ``;
+  showTasks(getRandomNumber(1, 10));
+};
+
+
+const templateTask = `<article class="card card--edit card--black">
+            <form class="card__form" method="get">
+              <div class="card__inner">
+                <div class="card__control">
+                  <button type="button" class="card__btn card__btn--edit">
+                    edit
+                  </button>
+                  <button type="button" class="card__btn card__btn--archive">
+                    archive
+                  </button>
+                  <button
+                    type="button"
+                    class="card__btn card__btn--favorites card__btn--disabled"
+                  >
+                    favorites
+                  </button>
+                </div>
+
+                <div class="card__color-bar">
+                  <svg width="100%" height="10">
+                    <use xlink:href="#wave"></use>
+                  </svg>
+                </div>
+
+                <div class="card__textarea-wrap">
+                  <label>
+                    <textarea
+                      class="card__text"
+                      placeholder="Start typing your text here..."
+                      filterName="text"
+                    >
+This is example of new task, you can add picture, set date and time, add tags.</textarea
+                    >
+                  </label>
+                </div>
+
+                <div class="card__settings">
+                  <div class="card__details">
+                    <div class="card__dates">
+                      <button class="card__date-deadline-toggle" type="button">
+                        date: <span class="card__date-status">no</span>
+                      </button>
+
+                      <fieldset class="card__date-deadline" disabled>
+                        <label class="card__input-deadline-wrap">
+                          <input
+                            class="card__date"
+                            type="text"
+                            placeholder="23 September"
+                            filterName="date"
+                          />
+                        </label>
+                        <label class="card__input-deadline-wrap">
+                          <input
+                            class="card__time"
+                            type="text"
+                            placeholder="11:15 PM"
+                            filterName="time"
+                          />
+                        </label>
+                      </fieldset>
+
+                      <button class="card__repeat-toggle" type="button">
+                        repeat:<span class="card__repeat-status">no</span>
+                      </button>
+
+                      <fieldset class="card__repeat-days" disabled>
+                        <div class="card__repeat-days-inner">
+                          <input
+                            class="visually-hidden card__repeat-day-input"
+                            type="checkbox"
+                            id="repeat-mo-1"
+                            filterName="repeat"
+                            value="mo"
+                          />
+                          <label class="card__repeat-day" for="repeat-mo-1"
+                            >mo</label
+                          >
+                          <input
+                            class="visually-hidden card__repeat-day-input"
+                            type="checkbox"
+                            id="repeat-tu-1"
+                            filterName="repeat"
+                            value="tu"
+                            checked
+                          />
+                          <label class="card__repeat-day" for="repeat-tu-1"
+                            >tu</label
+                          >
+                          <input
+                            class="visually-hidden card__repeat-day-input"
+                            type="checkbox"
+                            id="repeat-we-1"
+                            filterName="repeat"
+                            value="we"
+                          />
+                          <label class="card__repeat-day" for="repeat-we-1"
+                            >we</label
+                          >
+                          <input
+                            class="visually-hidden card__repeat-day-input"
+                            type="checkbox"
+                            id="repeat-th-1"
+                            filterName="repeat"
+                            value="th"
+                          />
+                          <label class="card__repeat-day" for="repeat-th-1"
+                            >th</label
+                          >
+                          <input
+                            class="visually-hidden card__repeat-day-input"
+                            type="checkbox"
+                            id="repeat-fr-1"
+                            filterName="repeat"
+                            value="fr"
+                            checked
+                          />
+                          <label class="card__repeat-day" for="repeat-fr-1"
+                            >fr</label
+                          >
+                          <input
+                            class="visually-hidden card__repeat-day-input"
+                            type="checkbox"
+                            filterName="repeat"
+                            value="sa"
+                            id="repeat-sa-1"
+                          />
+                          <label class="card__repeat-day" for="repeat-sa-1"
+                            >sa</label
+                          >
+                          <input
+                            class="visually-hidden card__repeat-day-input"
+                            type="checkbox"
+                            id="repeat-su-1"
+                            filterName="repeat"
+                            value="su"
+                            checked
+                          />
+                          <label class="card__repeat-day" for="repeat-su-1"
+                            >su</label
+                          >
+                        </div>
+                      </fieldset>
+                    </div>
+
+                    <div class="card__hashtag">
+                      <div class="card__hashtag-list"></div>
+
+                      <label>
+                        <input
+                          type="text"
+                          class="card__hashtag-input"
+                          filterName="hashtag-input"
+                          placeholder="Type new hashtag here"
+                        />
+                      </label>
+                    </div>
+                  </div>
+
+                  <label class="card__img-wrap card__img-wrap--empty">
+                    <input
+                      type="file"
+                      class="card__img-input visually-hidden"
+                      filterName="img"
+                    />
+                    <img
+                      src="img/add-photo.svg"
+                      alt="task picture"
+                      class="card__img"
+                    />
+                  </label>
+
+                  <div class="card__colors-inner">
+                    <h3 class="card__colors-title">Color</h3>
+                    <div class="card__colors-wrap">
+                      <input
+                        type="radio"
+                        id="color-black-1"
+                        class="card__color-input card__color-input--black visually-hidden"
+                        filterName="color"
+                        value="black"
+                        checked
+                      />
+                      <label
+                        for="color-black-1"
+                        class="card__color card__color--black"
+                        >black</label
+                      >
+                      <input
+                        type="radio"
+                        id="color-yellow-1"
+                        class="card__color-input card__color-input--yellow visually-hidden"
+                        filterName="color"
+                        value="yellow"
+                      />
+                      <label
+                        for="color-yellow-1"
+                        class="card__color card__color--yellow"
+                        >yellow</label
+                      >
+                      <input
+                        type="radio"
+                        id="color-blue-1"
+                        class="card__color-input card__color-input--blue visually-hidden"
+                        filterName="color"
+                        value="blue"
+                      />
+                      <label
+                        for="color-blue-1"
+                        class="card__color card__color--blue"
+                        >blue</label
+                      >
+                      <input
+                        type="radio"
+                        id="color-green-1"
+                        class="card__color-input card__color-input--green visually-hidden"
+                        filterName="color"
+                        value="green"
+                      />
+                      <label
+                        for="color-green-1"
+                        class="card__color card__color--green"
+                        >green</label
+                      >
+                      <input
+                        type="radio"
+                        id="color-pink-1"
+                        class="card__color-input card__color-input--pink visually-hidden"
+                        filterName="color"
+                        value="pink"
+                      />
+                      <label
+                        for="color-pink-1"
+                        class="card__color card__color--pink"
+                        >pink</label
+                      >
+                    </div>
+                  </div>
+                </div>
+
+                <div class="card__status-btns">
+                  <button class="card__save" type="submit">save</button>
+                  <button class="card__delete" type="button">delete</button>
+                </div>
+              </div>
+            </form>
+          </article>`;
