+++ src/utils.js
@@ -1,0 +1,3 @@
+export default function getRandomNumber(min = 0, max) {
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+}
